Revert "Merge pull request #28 from k-mx/master"
The markup effects the github pages site and is useful. As long as we
don't find a way to make it look pretty in both, the repo and the
github pages I'll prefer the looks of that pages site over the repo.

This reverts commit d554c41f5d891cc70e444314b05e3243e85c4547, reversing
changes made to beea6e7f9b88d40577d9523caf0242ac3a180731.